backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
django-environ==0.12.0